Handle different channels for admin and users
